+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add dismissedRequestEditorNotices to general settings. (`22551faa`)
 - Refactor release workflows for improved orchestration and versioning. (`c3291407`)
 
 ## 2.8.1 - 2026-07-26